Attach Pudding to window if in browser
This was present in the last version of Pudding; not sure whether
it was removed accidentally or deliberately, but having the Pudding
global available in the browser is surely an assumption made by
lots of users?

I've also updated the package.js that Meteor uses to point at
the built version.